test(pdf): extract figure rasterization tests
